+++ jwst/resample/resample_step.py
@@ -40,8 +40,8 @@ class ResampleStep(Step):
         enable_ctx = boolean(default=True)  # Compute and report the context array
         enable_err = boolean(default=True)  # Compute and report the err array
         report_var = boolean(default=True)  # Report the variance array
-        pixmap_stepsize = integer(default=1)  # Interpolation step size for pixel map; interpolation is used for stepsize > 1
-        pixmap_order = integer(default=1)  # Order of the pixel mapping polynomial fit, must be 1 or 3
+        pixmap_stepsize = float(default=1.0)  # Interpolation step size for pixel map; interpolation is used for stepsize > 1
+        pixmap_order = integer(default=1)  # Spline order for pixel mapping, must be 1 or 3
     """  # noqa: E501
 
     reference_file_types: list = []
